Remove Session and ProcessTrackingOptions
With the move toward helpers, such as the ExportMachine and
UniversalExporter structs, we do not need the lower level deciding how
to represent machine state.

The higher level helpers can implement the state recording as they
desire, instead of relying upon a pre-built state model. This affects
the perf_switch example, as it now has to manage it's own state. This
example could be moved over to the ExportMachine once the ExportMachine
can expose out the raw cswitch events if needed.

In the future, we expect many tools to either build upon ExportMachine
or UniversalExporter instead of the raw lower level pipelines. For tools
that use the lower level pipelines, they likely want to have full control
anyways around state management.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>