Event: Move to unified OS documentation
We need to ensure that documentation can be done for all of our code
regardless of the supported OS it is built on. We also want to have a
consistent way to define an OS extension, but have it be documented
without changing the structures.

Move Event toward a extension trait that is per-OS to ensure they can be
compiled anywhere it is supported and still have full documentation.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>