Unify OS documentation
We need to be able to build documentation on all OS's we support. There
are a few root level cfg's per-OS that need to get fixed.

Have openat, perf_event, and tracefs build when doc is defined,
regardless of OS. Fix up namespaces that are per-OS enough to make the
doc generation happy.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>