Exporting: Move ExportMachine to OS trait extension
The ExportMachine is used on both Windows and Linux. Currently we define
the implementation of parts of ExportMachine within each OS, which is
fragile. For example, it's possible to have methods that only exist on
Linux but not on Windows that are implemented for the ExportMachine.
Since this is done without trait extensions, it will only get documented
for that OS and lead to confusion.

Define common traits for ExportMachine that force equal implementations
on all OS's. This makes it clear that all OS's must support these
actions and cleans up the UniversalExporter symbol flows.

There are still per-OS methods that we need on ExportMachine or related
structs. Those need to still move to an extension trait. This way the
extension traits can be documented under the os namespace it's meant
for. Doing so aligns with how std has done per-OS extensions.

Future changes will address the remaining structs that still have dual
implementations in each OS. This change is focused on ExportMachine
specifically.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>