Exporting: Add drop closures
Exporting sometimes requires cleanup closures. Sometimes this can occur
after parsing the session and getting a new ExportMachine. Other times
we need to cleanup once the ExportMachine gets dropped. An example of
this is cleaning up local symbols. We only want to do that after there
is no more need for them.

Add methods to register closures to run once the ExportMachine gets
dropped. Expose these methods in the UniveralExporter for helpers to
use for hooks later.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>