Implement DotNet Ready To Run Symbol Resolution (#108)
* Expose the unwind algorithm for mappings.

During call stack capture, the stack is unwound using one or more
unwinding algorithms. Allows each mapping to determine the right unwind
algorithm to be used for the executable code contained therein.

* Implement symbol resolution for .NET ready to run pre-compiled code.

This support assumes that the r2rmap file that was generated at build
time is present next to the binary. The resolution algorithm is similar
to native code resolution, but uses an input r2rmap file instead of a
PDB file.

* Read the text VA and file offsets from the PE binary and use them to calculate
the base address from the beginning of the executable mapping that contains the
R2R code. This mimics the algorithm used to resolve symbols on Windows
without having to know where in memory the PE header was mapped in the
loaded layout of the ready to run PE image.

* Fix tests.

* Fix benchmarks.

* Add safety note.

---------

Co-authored-by: Brian Robbins <brianrob@linux.microsoft.com>