Exporting: Add CSwitches for Windows
The ExportMachine supports cswitches, and already does so for Linux.
However, Windows does not yet support this. We need this parity in order
to allow offcpu profiling on both Windows and Linux.

Integrate CSwitch data for Windows from ETW sources.

Move CSwitch HashSet out from ExportMachine to OSExportMachine to allow each
OS to save the appropriate details for cswitches (Windows needs to track a
PID from a callstack).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>