Exporting: Add os attributes and errors for replay
Formats need to be able to access the qpc frequency and CPU sampling
frequency both in and out of tree. We need to enable public methods for
accessing these.

Formats also need to have immutable access to the ExportMachine,
however, during replay we have a mutable reference. We need to break
this out for callers.

Add public accessors for ExportSettings, os_cpu_count, and os_qpc_freq
to ExportMachine.

Change the replay_by_time callback to include an immutable reference to
the ExportMachine to allow kind and string lookups.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>