Introduce MetricValue and ExportGraphMetricValueConverter.
MetricValue is responsible for representing the type of the metric
associated with a sample.  The type of metric is likely to be needed by
file format exporters so that they know how to convert the value to
something that makes sense for their file format.

ExportGraphMetricValueConverter defines the contract used by ExportGraph
to convert instances of MetricValue associated with samples into u64 to
be consumed by ExportGraph.  Users of ExportGraph can specify a
converter at creation time or simplify specify None and the default
converter, which is a straight pass through, will be chosen.