Exporting: Add NetTrace (without stacks)
This is the initial support for the existing NetTrace (1) format.

We have use a provider named "Universal" that has two sub-systems. The
first is the System subsystem which represents the process
reconstruction (process create, exit, mappings, symbols). The second is
the Event subsystem which represents the events that were collected and
their metadata. Each event is linked to a process from the System
subsystem.

In this change list I've only added the event metadata, in a later
changelist I will add writing them. However, this change list is enough
to validate that we can get machine state reconstruction done correctly
on both Windows and Linux with multiple processes being represented.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

1. https://github.com/microsoft/perfview/blob/main/src/TraceEvent/EventPipe/EventPipeFormat.md