ETW: Handle rundown providers
Some providers need to run at the very end of the collection session. On
Windows/ETW these are called rundown providers. We need a way for events
to easily describe this so we can automate when these get enabled and
any other rundown logic.

Add add_rundown_event() to EtwSession to make it easy to add an event
that should only have it's provider enabled at the end of the session.

Add add_rundown_callback() to EtwSession to make it possible for rundown
logic to be performed before the trace session is entirely stopped. An
example of this is enable live symbol collection for DotNet. Symbols
come in at the end and we need to wait until no more messages come in
for a period of time before stopping the trace.

Add flush_trace() to SessionCallbackContext so rundown callbacks can
flush the trace periodically to ensure newer events have been processed
before making a call to stop or not.

Fix bug that caused filters to be used incorrectly for event IDs if they
were empty (and also properly populate them).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>