Exporting: Add dynamic symbol concepts and hooks
Helpers need the ability to read and advertise dynamic symbols. An
example is DotNet JIT compiled symbols. These come in from ETW events on
Windows when methods are either compiled on the fly or when the trace is
being stopped via a rundown mechanism for existing processes.

In order to make this work, we need to have a way for helpers to hook up
their events directly to the underlying ExportMachine. Add
with_export_hook() to the UniversalExporter to allow helpers to clone a
mutable copy of the ExportMachine to publish dynamic symbols into.

Add add_dynamic_symbol() to ExportProcess which ensures helpers can
publish dynamic symbols efficiently. This uses a new ExportTimeSymbol
which is an ExportSymbol with a time attached. This allows dynamic
symbols to find the right mapping after the tracing is done, instead of
having to scan mappings on each publish.

Add os_add_dynamic_symbol() to each OS half of ExportMachine. This
ensures that each OS can do different logic (such as PID mapping).

Add needs_dynamic_symbol() to ExportProcess which ensures that if
symbols are from rundowns/process exits, we only add them if there are
IPs that used that symbol.

Add add_dynamic_symbol_mappings() to ExportProcess which is simply a
stub (will finish in later commits) that can create anonymous mappings
on the fly for symbols if we don't already have them (like on Windows).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>