Exporting: Implement dynamic mappings
We need to be able to check if a dynamic mapping already exists or
create one for a set of dynamic symbols. We also need to be able to
check if a dynamic symbol is even needed at the end of a collection.

Add SymbolPageMap to symbols.rs which allows marking and checking on
page boundaries (power of 2) if an IP has been seen and if a range ever
had an IP within it.

Utilize SymbolPageMap within ExportProcess to check if a dynamic symbol
is needed when the SYM_FLAG_MUST_MATCH is used.

Create dynamic ExportMappings for dynamic symbols that don't match
within existing mappings of a process. Create mappings on 4KB page
boundaries, it's very unlikely processes mix native compiled code and
dynamic code within the same page. We do not attempt to merge further
than that, so unique IPs within each 4KB page will have their own
ExportMapping. If this becomes an issue, we can revisit this behavior,
there is nothing preventing us from changing it later, callers are not
aware of this, etc.

Add self tests to ensure this functionality works as expected.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>