Implement user_event pre-registration. (#126)
Without this functionality, the process emitting the events must exist
and have pre-registered the events before the consumer attempts to
enable them. Failure to do so will result in the consumer encountering
failures when searching the system for the desired events.

With this functionality, the consumer pre-registers the desired
user_events definitions and then eneables them. When a process starts and
registers the matching events, the system will de-duplicate them, and
immediately enable the requested events.

Co-authored-by: Brian Robbins <brianrob@linux.microsoft.com>