user_events: Add FD Socket support
user_events is managed via a FD on Linux. This enables programs to
register events as long as they have an FD to
/sys/kernel/tracing/user_events_data. However, many programs do not have
direct access to tracefs. We can work around this by giving a FD
directly to user programs via UnixStreams. We plan to allow for this in
.NET via .NET's Diagnostic UnixStream endpoint.

Add WithUserEventFD public trait within the user_events module that
allow for sending data along with a user_events_data FD opened from
tracefs.

Implement WithUserEventFD for UnixStream for future use with .NET's
diagnostic endpoint.

Tested locally with a user program receiving the FD and registering an
event with it. Hard to make a self test for this.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>