NetTrace: Fix PID/TID 0 edge case
If a real sample comes in with a PID and TID of 0, then it must have an
ID of 0 within the nettrace format. Before this case could get a
non-zero ID and the non-zero ID would get removed subtly via the final
insert of the reserved entry before draining.

Instead of adding before draining, add a PID and TID of 0 as ID 0 into
the hashmap before it is used each time. This ensures if the PID and TID
0 case comes in from a real sample (on Linux it does), then it will
properly get ID of 0.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>