NetTrace: Deduplicate mapping metadata by ExportDevNode
A lot of size in the trace is taken up by replicating the mapping
metadata details. Instead of repeating this data, we should write this
data out on first use and refer to it by ID the next time it is seen.

Track mapping metadata by ExportDevNode during mapping writes. Only
actually write out a new mapping metadata event if this is the first
time we've seen it. Otherwise, just refer to it by ID.

Use ID 0 as a reserved ID that means "Not defined / Unknown" for
mappings that don't have a ExportDevNode (such as anonymous code
sections).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>