Exporting: Add record data concept to samples
Sometimes a sample can be as simple as an IP or Callstack along with a
value. Other times, the sample also needs more data, such as encoded
binary data that describes more details about the sample. An example is
dotnet's GCAllocationTick event, where details like the instance ID and
type name of the allocation are stored. These details are useful in low
level investigations, and formats that support them, need to be able to
export them as they were originally captured with minimal processing.
While some of these could be done via simple attributes, many of these
cannot be, as sometimes the collection source does not know the actual
format or care to parse them out.

To ensure we don't expand all ExportProcessSamples size, this change
introduces a level of indirection. The ExportProcessSample will point to
a record within the ExportMachine indicating where (if any) data is
stored and what format it is in. Add sample_record_data() to ExportMachine
that does this matching location logic for callers. Format/tools authors
can use this new method when they support high fidelity record data from
samples. Both the raw data and the format (fields, name, id) are
returned.

Add ExportRecordType and ExportRecordData structs that define what the
record data is (as originally collected, like field names) and where the
data is saved in a global vector (to prevent many allocations, we use a
shared vector for data across all processes). The ExportRecordType is
linked to a kind as well, this allows formats to support egressing data
that has the same sample kinds, but some have verbose data within them,
while others may not.

Add add_sample_with_record() to the ExportSampler and matching structs.
This allows very easy linking of extra (unparsed) binary data into a
sample for formats that support it (like NetTrace). For formats that
don't support it, they will simply ignore it, but still export out the
sample along with the callstack and value associated. This ensures we
can have high fidelity sources that are not limited by low fidelity
export formats.

Add set_record_type() to the ExportBuiltContext, which allows a single
time setup for callers to state what format the record data will be in
on a per-event basis. Use this data during add_sample_with_record() to
link the record data to the record type (Fields, Name, ID, etc).

Store record_id within ExportProcessSample that links to the global
record data table within ExportMachine. We don't expect more than high
millions of events per-ExportMachine, so we can limit callstack_id and
record_id down to u32's to save memory while externally treating it as a
usize during set/get operations.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>