NetTrace: Support sample data records
Samples can now have data records attached to them to describe extra
data, such as the original payloads of the events underpinning the
sample. NetTrace supports certain types of data in V6, which we want to
support matching fields for. This allows data records to show up in
tools that understand the NetTrace format, allowing for further
diagnostics when investigating traces.

Add method to convert between EventField type names, such as those found
on Linux, and NetTrace types. Stop once we come to a field we don't
understand (or there are no more fields). If we stop early, do not fail,
simply end the descriptions of the fields. The entire payload will still
get appended no matter what, however, there won't be nice descriptions
of it.

Make metadata entries for each record type that is within the
ExportMachine from events. This ensures NetTrace can determine if it's a
simple Sample or a complex Sample with extra data. Use the record type
name as the event name. Collection tools should make sure these names
are unique enough for reading cases. The actual event ID is currently
not saved anywhere, in the future we could put this into further
metadata if needed. The main reason for this is that event IDs cannot
conflict in NetTrace's metadata section per-provider name. Since all
events use the Universal.Events provider name, each event must have a
unique ID. We simply dynamically allocate a unique ID for each metadata
entry to ensure uniqueness.

Append record data from Samples with them to the NetTrace event entries.
Always append the full data, even if the description of it was truncated
due to unknown types, etc.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>