Add the text section VA offset to serialized PE metadata.
This is required to calculate RVAs when doing remote symbol lookup for
.NET ready to run images for processes running on Linux. CoreCLR load
behavior results in a single mapping being reported for the text
section, but RVAs must be calculated from the ImageBase virtual address.
The offset is used to translate from the text section virtual address to
the ImageBase virtual address and then from IP to RVA.

   RVA = IP - ExecutableMappingStartVA - text_offset

This can change if CoreCLR changes its load behavior. I intend to look
at adding tests to the dotnet/runtime repo to catch if this behavior
changes.