Scripting: Add scripting framework
Tools, such as record-trace, will require flexibility defining exactly
what to collect. This requires logic that is not easily put into a
direct commandline argument, such as do this on Windows vs do this on
Linux. Or enabling entire helpers within one-collect with per-helper
parameters.

The scripting framework utilizes Rhai and allow for trait extensions to
plumb types/objects and methods out to scripts. This allows a lot of
flexibility for both maintenance/freedom of helpers to do what they
require for users as well as not make each tool re-implement this logic.

Add ScriptedUniversalExporter struct that will act as the main place to
focus script extensions. Add appropriate hooks so other helpers can
extend the script support out to tools that wish to use it. Tools will
still have to register helpers with the exporter, but they will not have
to worry about the details. This ensures that tool owners can choose to
limit scripting support, if needed.

Add scripting feature to our crate. This ensures tool owners that do not
want any scripting support do not take on any of this code of complexity
and instead will rely entirely on CLI arguments.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>