Scripting: Add scripting support from one-collect
Add the DotNet supported scripting additions and basic engine from
one-collect to record-trace.

Add --script and --script-file arguments that allow callers to pass
either a snippet or file containing Rhai syntax.

Allow record-trace to run if a script is passed and on-cpu and off-cpu
are both turned off (IE: Event only collection cases).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>