Exporting: Ensure Data and Attributes can be read
Data from the event payload needs to be read and recorded in some cases.
The borrow checker rules prevent outside scopes from doing this cleanly.

Add add_sample_with_event_data() that takes a Range to make it possible
to add a sample with a record that originates from the event data.

Ensure OS attributes for events are read before taking the exporter
borrow. On Windows some attributes are pulled from the exporter and
cause a borrow violation.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>