Universal: Allow adding Events
Events sometimes need to be added to the ExportMachine. The
UniversalExporter already contains an ExportSetting where events are
typically added. However, we often want to be able to extend the
UniversalExporter to include more events after the ExportSettings have
been set.

Add add_event() method to UniversalExporter that takes the current
settings and adds an event to it. This allows the flexibility we need to
ensure we can add Events before or after the UniversalExporter has been
created.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>