DotNet: Establish Scripted Event Pattern
For Windows and Linux DotNet will emit events out via the runtime
provider. On Windows this is via ETW, on Linux this is via user_events.
Windows is already established, so we can use those to start with. We
need to be able to tie the script methods we expose to actual events and
provider/keyword/level values for ETW.

Add DotNetSample struct which describes the event details as well as how
to take the sample value. Most events will simply emit out a Count
sample of 1. However, events like GCAllocationTick, will emit out of a
Bytes metric with the actual allocation amount.

Add runtime_samples() to return the enabled events (via DotNetSample
structs) so they can be enabled within the OS specific providers.

Hookup Windows side of DotNetScenario to ETW via the runtime_samples()
method that we now have. As more providers are added, they will follow
the same pattern with a different provider GUID. Linux will use this
same pattern, but will have a "trampoline" event that multiplexes the
events. This is why DotNetSample sample_value closures only take a byte
array, as they will come from different places on Linux that need to be
specified at the OS hook level.

Add contention events to runtime so they can be accessed as well.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>