Event: Add dynamic field value closure support
Events describe fields via the EventFormat struct within it. Field
values are almost always needed to parse data out. As a part of
scripting Event support, we need a way for users to describe which field
to use for certain things, such as a sample value. This is
straight forward when there are no dynamic length fields (such as on
Linux tracepoints). However, when random length strings exist that are
not put behind a __rel_loc / __data_loc value, the offset of the next
fields is not clear at all. This requires readers to read the dynamic
length values and add their lengths to the offset of the trailing
fields. Let's standardize how to do this in a way that tools can use
this directly as well as scripts.

Add try_get_field_data_closure() to Event that returns a closure that
gets the data bytes (but not the type conversion) from a field name.
This simplifies these scenarios drastically as the closure can then be
used elsewhere (like other closures) to get the field values very
quickly without additional parsing.

All offsets and skips are pre-calculated outside of the closure, so
performance remains as good as hand coded routines for these types of
data. When there are no dynamic length values the closure turns directly
into a method call for the underlying data at a known offset.

Add self tests to ensure the skipping logic and variable length cases
functional properly.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>