Exporting: Add dynamic metric value support
Metric values can be several types (Count, Bytes, Duration) and can come
from various data types (u8, u16, u32, u64, etc.) Scripts will want to
describe what field to sample upon. The field maybe any type, so we need
a way to determine if we can process a given type to a metric and what
type of metric should it be.

Add try_get_value_closure() to MetricValue that takes a a type of metric
and a type of data. Using both it will determine if a closure can be
created. If it can, then the closure will return back valid data if
possible. Otherwise, it will return an error just as if a hand written
event callback would.

If the type of metric or type of data is unknown, then no closure is
returned. Scripts typically will error upon this condition instead of
defaulting to a default value.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>