Exporting: Script sampling and recording events
We will enable various ways to get events to sample and record via
scripts. We do not want to tie how you get an event to what you do with
the event. This is why the ScriptEvent struct exists. However, we do not
yet have a way to add a ScriptEvent to the ExportMachine.

Add sample_event() and record_event() script methods. These take a
ScriptEvent along with other parameters that describe what to do. When
sampling, the caller describes what type of metric to emit and where to
get that value. When recording, there are no options currently.

NOTE: Rhai allows function overloading, so it's possible to add
default-values for new parameters we want to expose without a new script
method. We also could just create another script method name if needed.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>