Exporting: Add Proxy Event Support
Sometimes events encode more than one event within their payload. An
example of this is EventHeader that encodings many events with a
payload containing first the event ID, then the event data following it.
The event ID describes the actual event ID to use instead of the
original event ID. We need a pattern where an event closure can parse
this data, then redirect at the ExportMachine level (lower level support
is not implemented, but could be done).

Add EVENT_FLAG_PROXY to Event and expose it out via methods. This can be
used by components that want to make an event that represents an event
that is encapsulated within a single parent event.

Add ExportProxy struct that takes events with the EVENT_FLAG_PROXY and
enables redirection lookups by ID. Expose this out via ExportSettings
and ensure a unique ID is given for each proxy event to enable lookup.

Add proxy_data() and proxy_event_data() to ExportTraceContext to allow
for ExportSettings::with_event() closures to proxy data to the correct
proxy event by the proxy event's ID (closure owns the lookup between the
event payload ID and the actual proxy ID).

Add self test (only on Windows due to PerfSession creation requirements)
to ensure proxy data is working as expected.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>