DotNet: Use OSDotNetEventFactory for all script events
DotNetScenarios make the underlying DotNet event directly via
Event::new(). This doesn't allow for an OS, like Linux, that uses a
proxy event to work correctly.

Have DotNetScenarios (and all other DotNetScript methods) use a single
instance of the OSDotNetEventFactory to ensure we can proxy (or do
whatever we need) on a per-OS basis.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>