Event: Add reading __rel_loc data offset
Fields that have __rel_loc data types need to get the offset and size
details to be useful. In scenarios where the data is not meant to be
read, but simply referred to, we need to get a Range for this value
instead of the actual data.

Add get_rel_loc() method to EventFormat that will return the Range for
the data that field points to (but not the actual data).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>