DotNet: Add Linux Tracepoint creation
DotNet events on Linux use an IPC mechanism (not implemented yet) that
requires a user_event to exist for events to be written there.

Add user_event creation of each provider and callstack pairs (IE: One
tracepoint for provider with no callstack data, and another with
callstack data). This is needed to ensure we can tell perf_events to turn
on callstacks or not at that level due to the multiplexing that DotNet
runtime implements.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>