DotNet: Enable Tracking UserEvent by Process
User_events via DotNet require sending data to the diagnostic socket of
a dotnet program. This requires tracking which processes are hosting
DotNet, and then knowing which events to use. We have the events to use
already, but we need to split them up by tracepoint name, keyword, and
level per-provider.

Add UserEventTrackerSettings which can describe the tracepoint name and
matching events along with the keyword and level to use.

Add UserEventTracepointEvents and UserEventProviderEvents for storing
the required details for the UserEventTrackerSettings.

Add UserEventTracker which starts a worker thread and enables processes
as they become active that host DotNet libraries.

Wire up UserEventTracker into the underlying session's mmap event if the
mmap is for a DotNet library.

Later changes will invoke the actual IPC commands, for now, we just open
up the diagnostic socket and keep it open until the session is stopped.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>