Exporting: Add timeline scripting support
Often times tools want to be able to record a timeline of several events
together that share a common key, such as a request ID. An example of
this on Linux is the block IO events. Being able to see when a given IO
was inserted into the queue, issues to the underlying device,
potentially requeued, and completed is incredibly useful.

Add ScriptTimeline struct and matching Rhai methods and types for
scripts to describe a timeline. Scripts invoke new_timeline() to get a
ScriptTimeline and use with_event() to describe events that are either
intermediate or completions along with the field(s) that make up the key
linking them together. Allow up to 4 keys, as long as the sum data is
32-bytes of less.

Add ExportTimeline struct that actually handles the closure logic and ID
key correlations. This allows both scripts and other tools to re-use
this logic easily.

Add set_min_*() methods to Script/ExportTimeline to allow describing to
only record/save the data if the end-to-end time exceeds a given value.
This ensures we don't have data files that export short/normal periods
when we only want certain longer periods.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>