Exporting: Add sample hooks and insulate adding
It's often useful to drop samples that don't make a certain criteria
holistically across the entire machine. It's also often useful to simply
print out all samples and not actually save ALL of them.

Add sample_hooks to ExportMachine to allow for both of these scenarios
by tools.

Insulate adding record_data to samples to only common pathways to ensure
that tools do not fall (easily) into patterns that result that record
data is saved when the sample was not interesting.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>