Record-trace: Add --live option
When data is being recorded sometimes it is useful to watch the data
come in. This allows engineers to see if anything interesting has been
recorded, and once it has, stop tracing further.

Print out Count, Bytes, and Duration values for samples seen if --live
is specified by the user. Callstacks and records are not printed.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>