Exporting: Protect kernel from comm/mmap changes
The PID 0 process is reserved for the kernel within the ExportMachine.
However, on Linux, PID 0 can be given for a process outside of the
current container's PID namespace. When this happens, we need to ensure
the kernel process comm and mmap details do not get modified.

Add comment and check for PID 0 in both add_comm_exec() and
add_mmap_exec() to prevent changes to the kernel process from read
events. If kernel specific manipulation is required, it can be done
directly via process_mut() instead of the events.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>