BUG: Fix infinite loop in BufRead.lines().flatten()
We have several places that use the pattern of BufRead.lines().flatten()
to easily read all the lines within a given file. However, if a line is
returned that had an error, flatten() will continue to ask for another
line. This causes an infinite loop on files that are no longer
accessible, such as procfs or tracefs.

Move all instances of flatten() to the proper match case that breaks out
of the loop if an error is hit.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>