Exporting: Custom samples need TID provided
When we make custom samples, we do not currently allow the TID to be
overriden. However, we allow the PID to be overriden via
add_pid_sample() or add_custom_sample().

Move all of our make_sample() functions to require a TID, update
callers. Use provided TID for the sample TID instead of the native OS.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>