Exporting: Support Spans as MetricValues
Often times tools need to capture a set of events that occur and
represent them in some way. Some formats will allow saving every event,
while otherse may only support the duration of time it took.

In order to support all formats, add a Span MetricValue entry that
points to the ID of a recorded ExportSpan.

Add ExportSpan struct that describes the start and end time of an
activity. Include the name and ensure children can also be saved.

Ensure all QPC duration operations are done within ExportMachine helpers
to ensure it is done properly and consistently.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>