Exporting: Add Spans to Timelines
Now that we have ExportSpan structs and the general Span concept, we
should make them for timelines. This allows users to see the various
steps in a timeline and how long each one took. In the future, we will
also allow payload recording of steps to allow for even further
analysis.

Change the ScriptedTimeline with_event() script method to take a
flag rather than just a bool. This allows for more control and future
growth of options without breaking scripts going forward. This will be a
one-time break, as no one is really using this yet.

Build ExportSpans when we apply a timeline using the new flags to derive
when an event is a start, end, or transient event. Start and transient
events act as child spans of the single timeline span. End event simply
ends the parent and last child span.

Export add_span() method out to export event closures via the
ExportSampler. This allows event closures to add a span and get a
MetricValue back that represents that span within the ExportMachine.
In the future, if an event is filtered out, we need to remove the span
from the set. However, that is not covered in this change.

Add self tests to ensure flag policies are working correctly for
timeline. Ensure basic functionality and error handling.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>