Exporting: Add Timeline record field support
Timelines are composed of several events. Typically only 1 or 2 fields
out of all those events want to be recorded. We need to allow for
describing which fields from which events should be used for the
Timeline Span value associated record. This will allow for tools to only
capture what they need and to combine different fields from several
events into a single record.

Add should_record_field() to TimelineEventFlags to describe one or more
fields that should be recorded from that Event.

Add sample_record_data() to ExportSampleFilterContext to allow for
filters to look at any associated record data to either live print it or
to scan it for content that should not be saved (such as HBI, PII, etc).

Add add_pid_sample_with_record() to ensure that events can add a sample
that has a custom record from another PID/TID pair.

Utilize record fields in the apply_timeline macro to properly record the
wanted fields. Last write wins with multiple Events both wanting the
same field overwriting each other. This is assumed to be rare, but
wanted to call out the behavior for our history.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>