Record-trace: Print associated sample records live
Samples that contain records aren't printed during live output. Often
times the record is important, such as how many Bytes were copied during
a single or composite event (such as a Timeline).

Utilize the new EventFormat::get_write_closure() to add the known fields
of a record to the output.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>