Event: Add Field Filter Closure Support
Often times it's useful to be able to filter out events by their field
values. An example is only wanting storage events that were over a
certain size. We already have a pattern for getting field closures, we
can leverage this to also have filters.

Add try_get_field_filter_closure() that utilizes the already existing
data closure mechanism to get the data of a field and filter on it.
Support both numeric and string comparisons.

The comparisons for numbers are ==, !=, <, <=, >, and >=.

The comparisons for strings are ==, !=, contains, not_contains,
starts_with, and ends_with. All string comparisons are case sensitive.

Add self tests to ensure filtering is properly handled for UTF8/16
strings as well as all supported numeric types
(u8/16/32/64/int/short/long).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>